requirements.txt
Flask-SocketIO~=5.0.1
Flask~=1.1.2
jsonschema~=3.2.0
PyYAML~=5.4.1
svarog~=0.1.2
